updates create_post function, intentional and automated formatting
